update google map shortcode
The map shortcode was completely broken because the callback function
was spelled wrong. That has been addressed (along w/ a couple unrelated
syntax errors).

Now a user has an additional param called `type`. If they enter
"address" in that param they can enter an address into the URL field and
it will be converted on google's side. This addresses the changes in the
google map embedding