Welcome page and Pro notice fixes
This fix enures that it will show pro plugins notices only when the WC is active ib the customers site.

Welcome page will be displyed only for admin users, other users it will be not dissplayed.

Now, if there are 2 or more than that our plguins are active on customers site then We will not disply welcome page. As it was causing the infinite loop of the welcome page.